+++ ufo/static/chromePolicy.html
@@ -1,3 +1,4 @@
+<link rel="import" href="bower_components/paper-button/paper-button.html" />
 <link rel="import" href="bower_components/paper-item/paper-item.html" />
 <link rel="import" href="bower_components/paper-listbox/paper-listbox.html" />>
